Fix index error in _parallel_processes
If the size of an array is smaller than the number of processes, only
create the number of tasks required rather than a task for every process

If shape[0] is 4 and _NUMBER_OF_WORKER_PROCESSES is 48, only 4 tasks
should be created, each with a slice of size 1.

Also fix the --multiprocessing flag so that it sets the global values
of _NUMBER_OF_WORKER_PROCESSES, rather than creating a local variable
of the same name.

See #400