Change _pearson_fit logical relation from `and` to `or`
One way _pearson_fit sanitizes the values it will be working
with is by setting any null or miniscule values to a predetermined
minimum value.

However, it checks for whether a value is null or miniscule by
only looking for values that are *both* null and miniscule. Change
this so that the logic works as intended.

See #421